refactor: inline src/ modules back into main.js — single file, no runtime require
